v0/requirements.txt
scipy
numpy
seaborn
networkx
antlr4-python3-runtime
pint
babel
